Adjust debug API to address some x86-64 concerns.
- The argument buffer contained in the debug_{pre,post}_syscall message structures wasn't large enough to accomodate all
arguments for some syscalls on x86-64, which could potentially have led to kernel memory corruption when using syscall
tracing via the debug API. As such, enlarge it to accomodate 64-bit platforms as well.

- Adjust TeamDebugger/SyscallInfo to discriminate the target architecture and read the arguments when trapping console
output. Gets the latter working on x86-64.

+++ src/apps/debugger/model/SyscallInfo.cpp
@@ -31,7 +31,7 @@ SyscallInfo::SyscallInfo(const SyscallInfo& other)
 
 
 SyscallInfo::SyscallInfo(bigtime_t startTime, bigtime_t endTime,
-	uint64 returnValue, uint32 syscall, const uint32* args)
+	uint64 returnValue, uint32 syscall, const uint8* args)
 	:
 	fStartTime(startTime),
 	fEndTime(endTime),
@@ -44,7 +44,7 @@ SyscallInfo::SyscallInfo(bigtime_t startTime, bigtime_t endTime,
 
 void
 SyscallInfo::SetTo(bigtime_t startTime, bigtime_t endTime, uint64 returnValue,
-	uint32 syscall, const uint32* args)
+	uint32 syscall, const uint8* args)
 {
 	fStartTime = startTime;
 	fEndTime = endTime;
